Upgrade ESLint configuration to new flat config
We sadly have to get rid of cypress for the moment.
