+++ package.json
@@ -19,8 +19,8 @@
   },
   "devDependencies": {
     "@stylistic/eslint-plugin": "^1.5.0",
-    "eslint": "^8.55.0",
-    "eslint-plugin-cypress": "^2.15.1",
-    "eslint-plugin-erb": "^1.1.0"
+    "eslint": "^9.0.0-alpha.0",
+    "eslint-plugin-erb": "^2.0.0",
+    "globals": "^13.24.0"
   }
 }
